requirements.txt
asgiref==3.11.1
boto3==1.42.96
botocore==1.42.96
Django==6.0.4
django-allauth==65.16.1
django-environ==0.13.0
django-mathfilters==1.0.0
django-storages==1.14.6
django-widget-tweaks==1.5.1
gunicorn==25.3.0
jmespath==1.1.0
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.16.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
